+++ moodle/tests/fixtures/moodle_php_forbiddentokens.php
@@ -9,4 +9,6 @@ label1:
 echo 'Here we are, thanks to lovely goto';
 exit(1);
 label2: echo 'Will you understand ever that goto is forbidden?';
+// Backticks aren't welcome as execution operators.
+$files = `ls -al`;
 // Fair enough.